open window

+++ current_lib/libft/ft_substr.c
@@ -19,9 +19,9 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 
 	if (!s)
 		return (NULL);
-	if (start >= ft_strlen(s))
+	if (start >= (size_t)ft_strlen(s))
 		return (ft_strdup(""));
-	if (start + len > ft_strlen(s))
+	if (start + len > (size_t)ft_strlen(s))
 		len = ft_strlen(s) - start;
 	sub = (char *)malloc((len + 1) * sizeof(char));
 	if (!sub)
